fixed bugs

+++ backend/services/annotation.py
@@ -4,20 +4,30 @@ from services.supabase import supabase_client
 annotation_table = supabase_client.table("annotations")
 class AnnotationService:
 
+    def annotation_already_exists(question_id: str, annotator_id: str): 
+        response = annotation_table.select("*").eq("annotator", annotator_id).eq("question_id", question_id).execute()
+        data = response.data 
+        return len(data) > 0
+
     @staticmethod
-    def publish_annotation(id, question_id, answer, time_spent, segments_watched, annotator):
+    def publish_annotation(id, question_id, answer, time_spent, segments_answered, segments_watched, annotator):
+        if annotation_already_exists(question_id, annotator_id):
+            return None
+
         annotation = Annotation(
             id=id,
             question_id=question_id,
             answer=answer,
             time_spent=time_spent,
+            segments_answered=segments_answered,
             segments_watched=segments_watched,
             annotator=annotator
         )
 
         response = annotation_table.insert(annotation.to_json()).execute()
         return response.data
-
+    
+        
     @staticmethod
     def get_complete_questions(annotator_id: str):
         response = annotation_table.select("question_id").eq("annotator", annotator_id).execute()
